+++ src/timer/slot.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 
 //Slot is based on HashMap, It easy to add it and find it.
 pub struct Slot {
-    task_map: HashMap<u32, Task>,
+    task_map: HashMap<usize, Task>,
 }
 
 impl Slot {
@@ -17,14 +17,14 @@ impl Slot {
         self.task_map.insert(task.task_id, task)
     }
 
-    pub(crate) fn remove_task(&mut self, task_id: u32) -> Option<Task> {
+    pub(crate) fn remove_task(&mut self, task_id: usize) -> Option<Task> {
         self.task_map.remove(&task_id)
     }
 
     //check并减cylinder_line，
     //返回现在要运行的，TaskOwned的集合
     //cylinder_line == 0
-    pub fn arrival_time_tasks(&mut self) -> Vec<u32> {
+    pub fn arrival_time_tasks(&mut self) -> Vec<usize> {
         let mut task_id_vec = vec![];
 
         for (_, task) in self.task_map.iter_mut() {
